% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARU_batch_import.r
\name{ARU_batch_import}
\alias{ARU_batch_import}
\title{A function to import a set of csv files.}
\usage{
ARU_batch_import(folderdir, FORMAT = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = "America/Bogota")
}
\arguments{
\item{folderdir}{is the folder containing the series of csv files you wish to import}

\item{FORMAT}{is a character string specifying the time stamp format 
(be sure to specify a time zone if working with data that covers multiple time zones or includes ST and DST)
provided the card was programmed using US date formating (mm/dd/YYYY) there should be no issue with the default}

\item{tz}{Specify the time zone you want to use. This is metadata that is stored internally, but will ease
transformations later on. The default is 'America/Bogata' which is GMT-5. This is set because our ARU's were initially
set to CDST (GMT-5) and left year round. The America/Bogtata time zone does not observe DST.}
}
\description{
produces a list where each frequency is a separate data frame and the tower the record came from is an
additional variable
}
\details{
This function assumes csv file names follow a pattern where they file name ends as 'TOWER_X.csv where
 X is a number 0-9
}
\examples{
a<-ARU_batch_import('C:/Users/mydirectory,FORMAT='\%Y-\%m-\%d \%H:\%M:\%S',tz='America/Bogota)
}
\keyword{ARU}

