% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARU_quickplot.r
\name{ARU_quickplot}
\alias{ARU_quickplot}
\title{Quickly visualize the ARU telemetry records}
\usage{
ARU_quickplot(batch_import_object, Freq, FROM = "", TO = "")
}
\arguments{
\item{batch_import_object}{The object created by ARU_batch_import}

\item{Freq}{The position or exact name of the frequency you want to view.}

\item{FROM}{Starting date to display data, mm/dd/yyyy}

\item{TO}{Ending date to display data, mm/dd/yyyy}
}
\description{
Plot ARU records for a single frequency between two dates, one plot for each tower available
This is a stand alone script, and should not be assigned to an object
}
\details{
Takes imported csv files of raw ARU data and converts them into a plot for a specific frequency
Signal strength records are rounded to the nearest 200db to better identify sedentary periods
Contaminated records (noise db<-13700) have not been removed
}
\examples{
names(batch_import_object) # obtain the positions and exact names for the different frequencies available
ARU_quickplot(mydata, 3,FROM="05/22/2013",TO="05/28/2013")
}
\keyword{ARUplot}

