% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indv_output.r
\name{Indv_output}
\alias{Indv_output}
\title{Individual output}
\usage{
Indv_output(mydata, outfolder, DATE = "")
}
\arguments{
\item{mydata}{object created by batch import}
}
\description{
complete later
}
\examples{
Indv_output()
}
\keyword{TBD}

