% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_worklist.r
\name{make_worklist}
\alias{make_worklist}
\title{A function to create  a worklist for an ARU manufactured by JDJC Corp. Fisher, IL}
\usage{
make_worklist(masterlist, outpath, name, Group_value = "", Group_var = "",
  numant = 6, ovrh = 0.2, SYNC = TRUE)
}
\arguments{
\item{masterlist}{the input table, see above for details about this table}

\item{outpath}{where the text file will be created}

\item{name}{the name of the setup file to be created}

\item{Group_value}{the level of Group_var to subset and select}

\item{Group_var}{a grouping variable that can be used to subset frequencies. If not needed, you still must create one and set all the values the same}

\item{numant}{the number of antennas. Needed when computing a SYNC time, default is 6}

\item{ovrh}{the overhead, default of 0.2. Only needed for computing a SYNC time}

\item{SYNC}{do you want a SYNC time appended to the start of the worklist? Defaults to TRUE}
}
\description{
Takes a data frame with named columns TunedFreq,pw,pint and a grouping column, at minimum
and produces a setup text file that can be loaded directly to an SD card. This is a stand-alone script, do not 
assign it to an object
}
\details{
This function requires the input table to have columns with the exact names TunedFreq,the frequency the ARU will
 look for, pw, the pulse width, and pint, the pulse interval
This function automatically appends the date the file was created to the file name
}
\examples{
make_worklist(mytable,outpath='C:/Users/myDesktop/',name='worklist',Group_var=Landscape,Group_value='SIB')
}
\keyword{worklist}

