% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARU_ouput_wide.r
\name{output_wide_aru}
\alias{output_wide_aru}
\title{Wide output for mike}
\usage{
output_wide_aru(importobject, outfolder, DATE = "", SITE = "")
}
\arguments{
\item{import}{object an object produced by batch import}
}
\description{
will this even be included in the future
}
\details{
finish documentation later
}
\examples{
output_wide_aru()
}
\keyword{TBD}

